refactor: split web-control-ui into two-layer architecture
- Extract control-mode-view from monolithic main.ts
- Add app-shell with mode switcher (USE/CONTROL)
- Create modular structure: core/, modes/, product/
- Preserve all existing features: chat, rollback, auth, message categorization
- Default to USE mode, explicit switch to CONTROL mode
