+++ apps/web-control-ui/index.html
@@ -6,7 +6,7 @@
     <title>OpenClaw Web Control UI</title>
   </head>
   <body>
-    <web-control-ui-app></web-control-ui-app>
+    <app-shell></app-shell>
     <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
